Split imports into update, imports, filters, & compat list
